Exact float to rationial conversion
This version of rational_in_float() converts the float into an
int64/int64 rational and calculates the int32/int32 rational with the
least error (where possible reproducing it exactly) using continued
fractions.

Additionally a funktion rational_limit_denominator() is definined
which finds the best approximation for a rational with the denominator
not exceeding a given value. One use case is to recover fractions that
are stored with limited precision:

rational_limit_denominator(0.3333::float, 1000) -> 1/3